Map trees within a quadrat.

Map trees within a quadrat.

ggplot_dbh_bubbles_by_quadrat(vft,
  title_quad = "Site Name, YYYY, Quadrat:",
  header = header_dbh_bubbles(), theme = theme_dbh_bubbles(),
  lim_min = 0, lim_max = 20, subquadrat_side = 5, tag_size = 2,
  move_edge = 0, status_d = "dead")

Arguments

vft
A dataframe -- specifically, a ForestGEO ViewFullTable.

title_quad
A string to use as a title.

header
A string to use as a header (see headers).

theme
An object of class theme.

lim_min, lim_max
Minimum and maximum limits of the plot area.

subquadrat_side
Length in meters of the side of a subquadrat.

tag_size
A number giving tag size. Passed to ggrepel::geom_text_repel.

move_edge
A number to adjust the extension of the grid lines beyond the plot limits.

status_d
A character string indicating the value of the variable status that corresponds to dead stems.

Value

A list which each element is a plot of class ggplot.

See also

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(fgeo.tool)

#> 
#> Attaching package: 'fgeo.tool'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following object is masked from 'package:stats':
#> 
#>     filter

# Using a small viewFullTable for example (see ?vft_1quad)
small_vft <- sample_n(vft_1quad, 50)

# Filter the data you want. For example:
want <- filter(small_vft, CensusID == 4, PlotID == 1)
p <- ggplot_dbh_bubbles_by_quadrat(want)

#> * Appending tags of dead trees with the suffix '.d'

#> * Standarizing `dbh` by the count of `dbh` measurements

#> Error in ggplot(.df, aes(x = qx, y = qy)): unused argument (aes(x = qx, y = qy))

# Visualizing only the first plot
p[[1]]

#> Error in eval(expr, envir, enclos): object 'p' not found

# Printing all plots to .pdf, with parameters optimized for size letter

# NOT RUN {
pdf("map.pdf", paper = "letter", height = 10.5, width = 8)
p
dev.off()
# }

# Be careful if filtering by DBH: You may unintentionally remove dead trees.
# You should explicietly inlcude missing `DBH` values with `is.na(DBH)`
p <- filter(small_vft, DBH > 20 | is.na(DBH))
ggplot_dbh_bubbles_by_quadrat(p)[[1]]

#> * Appending tags of dead trees with the suffix '.d'

#> * Standarizing `dbh` by the count of `dbh` measurements

#> Error in ggplot(.df, aes(x = qx, y = qy)): unused argument (aes(x = qx, y = qy))

# Customizing the maps ----------------------------------------------------
# A custom title and header
myheader <- paste(
  " ",
  "Head column 1                     Head column 2                          ",
  " ",
  " ........................................................................",
  " ........................................................................",
  sep = "\n"
)

ggplot_dbh_bubbles_by_quadrat(small_vft, title_quad = "My Site, 2018. Quad:", header = myheader)

#> * Appending tags of dead trees with the suffix '.d'

#> * Standarizing `dbh` by the count of `dbh` measurements

#> Error in ggplot(.df, aes(x = qx, y = qy)): unused argument (aes(x = qx, y = qy))

# Tweak the theme with ggplot
library(ggplot2)

#> 
#> Attaching package: 'ggplot2'

#> The following object is masked from 'package:fgeo.map':
#> 
#>     ggplot

ggplot_dbh_bubbles_by_quadrat(
  small_vft,
  title_quad = "My Site, 2018. Quad:",
  header = header_dbh_bubbles("spanish"),
  tag_size = 3,
  theme = theme_dbh_bubbles(
    axis.text = NULL,  # NULL shows axis.text; element_blank() doesn't.
    plot.title = element_text(size = 15),
    plot.subtitle = element_text(size = 5),
    panel.background = element_rect(fill = "grey")
  )
)

#> * Appending tags of dead trees with the suffix '.d'

#> * Standarizing `dbh` by the count of `dbh` measurements

#> Error in ggplot(.df, aes(x = qx, y = qy)): unused argument (aes(x = qx, y = qy))