Map species and elevation data.

These functions wrap a number of map elements for convenience:

Use ggplot_species_or_elevation() to map species and optionally elevation data in a single page. You can map multiple species on the same plot or you can facet the output to map each species on a single plot and all plots in a single page.

Use ggplot_elevation() if you want to map only elevation in the simplest way.

Use ggplot_each_species() to apply the function ggplot_species_or_elevation() to each species in a census dataset. The output is not a map but a list of maps, one per species, that can be printed on a .pdf file.

ggplot_species_or_elevation(census, elevation = NULL, fill = "black",
  shape = 21, facet = TRUE, point_size = 3, contour_size = 0.5,
  low = "blue", high = "red", hide_color_legend = FALSE,
  bins = NULL, add_elevation_labels = TRUE, label_size = 3,
  label_color = "grey", xyjust = 1, fontface = "italic",
  xlim = NULL, ylim = NULL, custom_theme = NULL)

ggplot_elevation(elevation, contour_size = 0.5, low = "blue",
  high = "red", hide_color_legend = FALSE, bins = NULL,
  add_elevation_labels = TRUE, label_size = 3, label_color = "grey",
  xyjust = 1, fontface = "italic", xlim = NULL, ylim = NULL,
  custom_theme = NULL)

Arguments

census
A dataframe; specifically, a ForestGEO's census.

elevation
A list or dataframe giving ForestGEO's elevation-data.

fill
Character; either a colour or "sp", which maps each species to a different color.

shape
A number giving point size and shape (as in graphics::points()). Passed to ggplot2::geom_point().

facet
(Not available for ggplot_each_species()) Logical; TRUE wraps multiple maps within the area of a single graphic plot.

point_size
A number giving point size and shape (as in graphics::points()). Passed to ggplot2::geom_point().

contour_size
A number giving the size of the contour of elevation lines. Passed to ggplot2::stat_contour() (see ggplot2::geom_contour()).

low
A string giving a color of the elevation lines representing low and high elevation.

high
A string giving a color of the elevation lines representing low and high elevation.

hide_color_legend
Logical; TRUE hides the color legend.

bins
A number giving the number of elevation lines to map.

add_elevation_labels
Logical. TRUE labels the elevation lines with text.

label_size
A number (label_size) or character string (label_color and fontface) giving the size, colour and fontface of the text labels for the elevation lines.

label_color
A number (label_size) or character string (label_color and fontface) giving the size, colour and fontface of the text labels for the elevation lines.

xyjust
A number to adjust the position of the text labels of the elevation lines.

fontface
A number (label_size) or character string (label_color and fontface) giving the size, colour and fontface of the text labels for the elevation lines.

xlim
A vector of length 2, for example c(0, 500), giving the minimum and maximum limits of the vertical and horizontal coordinates.

ylim
A vector of length 2, for example c(0, 500), giving the minimum and maximum limits of the vertical and horizontal coordinates.

custom_theme
A valid ggplot2::theme(). NULL uses the default theme theme_default().

Value

ggplot_each_species() returns a list of ggplots.

ggplot_elevation() and ggplot_species_or_elevation() return a ggplot.

Examples

# Small dataset with a few species for quick examples
some_sp <- c("PREMON", "CASARB")
census <- subset(fgeo.data::luquillo_tree5_random, sp %in% some_sp)
elevation <- fgeo.data::luquillo_elevation

# Simplest way to map elevation data only
ggplot_elevation(elevation)

#> Error in ggplot(data, aes(gx, gy, z = elev)): unused argument (aes(gx, gy, z = elev))

ggplot_species_or_elevation(census)

#> Error in ggplot(data, aes(gx, gy)): unused argument (aes(gx, gy))

ggplot_species_or_elevation(census, elevation)

#> Error in ggplot(data, aes(gx, gy, z = elev)): unused argument (aes(gx, gy, z = elev))

# For maximum control, you can compose maps as you like
# Traditional sintax: g(f(x))
add_elevation_contrours(ggplot_base_elevation(elevation))

#> Error in is_named(x): unused argument (aes(gx, gy, z = elev))

# With the pipe: f(x) %>% g()
ggplot_base_elevation(elevation) %>%
  add_elevation_contrours()

#> Error in ggplot(data, aes(gx, gy, z = elev)): unused argument (aes(gx, gy, z = elev))

# The traditional sintax is hard to read when you compose multiple functions.
# With the pipe, readability isn't affected by the number of functions.
ggplot_base_elevation(elevation) %>%
  axis_limits(xlim = c(-100, 400)) %>%
  add_elevation_contrours(contour_size = 0.5) %>%
  add_elevation_labels(label_color = "red") %>%
  hide_axis_labels() %>%
  hide_color_legend() %>%
  add_species(census, point_size = 5) %>%
  facet("sp") %>%
  theme_default(legend.position = "top")

#> Error in ggplot(data, aes(gx, gy, z = elev)): unused argument (aes(gx, gy, z = elev))