Map tree tags by status, showing four subquadrats per plot-page.

This function maps tree tags by status and outputs a list of maps that can be printed on a .pdf file. Each map shows four subquadrats within a quadrat. The symbols on the map represent the status of each tree -- not the status of each stem. Although you should likely provide data of only one or two censuses, ggplot_tag_status_by_subquadrat() will summarize the data to reduce overplotting. The data on the map summarizes the history of each stem across all censuses provided. Each tag will appear in the map only once or twice: * A tag will appear once if it belongs to a tree which status was unique across all censuses provided -- either "alive" or "dead". * A tag will appear twice if it belongs to a tree which status was "alive" in at least one census, and also "dead" in at least one other census. This feature avoids unintentional overplotting and makes interpreting the map easier.

ggplot_tag_status_by_subquadrat(vft, x_q = 20, x_sq = 5, y_q = 20,
  y_sq = 5, subquad_offset = NULL, bl = 1, br = 2, tr = 3,
  tl = 4, title_quad = "Site Name, YYYY. Quadrat:", show_page = TRUE,
  show_subquad = TRUE, point_shape = c(19, 4), point_size = 1.5,
  tag_size = 3, header = header_tag_status(),
  theme = theme_tag_status(), move_edge = 0)

Arguments

vft
A dataframe -- specifically, a ForestGEO ViewFullTable.

x_q, y_q
Size in meters of a quadrat's side. For ForestGEO sites, a common value is 20.

x_sq, y_sq
Size in meters of a subquadrat's side. For ForestGEO-CTFS sites, a common value is 5.

subquad_offset
NULL or -1. NULL defines the first column of subquadrats as 1. -1 defines the first column of subquadrats as 0.

subquad_offset = NULL    subquad_offset = -1
---------------------    -------------------
     14 24 34 44             04 14 24 34
     13 23 33 43             03 13 23 33
     12 22 32 42             02 12 22 32
     11 21 31 41             01 11 21 31

bl, br, tr, tl
Label each of the four maps of a quadrat. See paginate().

title_quad
A string to use as a title.

show_page
Logical; FALSE removes the page label from the map title.

show_subquad
Logical; FALSE removes subquadrat labels on each map.

point_shape
A vector of two numbers giving the shape of the points to plot (see possible shapes in the documentation of ?graphics::points(), under the section entitled 'pch' values).

point_size
A number giving points size. Passed to ggplot2::geom_point().

tag_size
A number giving tag size. Passed to ggrepel::geom_text_repel.

header
A string to use as a header (see headers).

theme
An object of class theme.

move_edge
A number to adjust the extension of the grid lines beyond the plot limits.

Value

A list of ggplots, where each element of the list is a map of tree tags by status, showing four subquadrats.

Acknowledgment

Useful ideas and guidance came from Suzanne Lao, Stuart Davis, Shameema Jafferjee Esufali, David Kenfack and Anudeep Singh. Anudeep Sinh also wrote the algorithm to calculate subquadrats.

See also

Examples

library(fgeo.map)
library(fgeo.tool)

# Filtering the data to map -----------------------------------------------
# Filter the data you want. Using a small ViewFullTable for example:
# (see `?vft_1quad`)
small_vft <- dplyr::sample_n(vft_1quad, 50)
want <- filter(small_vft, CensusID == 4, PlotID == 1)
p <- ggplot_tag_status_by_subquadrat(want)

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

# Visualizing only the first plot
p[[1]]

#> Error in eval(expr, envir, enclos): object 'p' not found

# Print all plots to .pdf, with parameters optimized for size letter

# NOT RUN {
pdf("map.pdf", paper = "letter", height = 10.5, width = 8)
p
dev.off()
# }

# Be careful if filtering by DBH: You may unintentionally remove dead trees.
# * If you filter by `DBH`, you loose the dead trees becaue their `DBH = NA`
# * You should explicietly inlcude missing DBH values with `is.na(DBH)`
p <- filter(vft_4quad, DBH > 20 | is.na(DBH))
ggplot_tag_status_by_subquadrat(p)[[1]]

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

# Customizing the maps ----------------------------------------------------
# Common tweaks
p <- ggplot_tag_status_by_subquadrat(small_vft, show_page = FALSE, show_subquad = FALSE)

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

p[[1]]

#>    [1] 384550 384553 384554 384555 384556 384557 384558 384560 384561 384562
#>   [11] 384563 384564 384570 384575 384576 384577 384579 384580 384582 384584
#>   [21] 384587 384589 384590 384591 384592 384593 384594 384595 384596 384597
#>   [31] 384599 384600 384601 384602 384607 384608 384610 384611 384612 384613
#>   [41] 384614 384615 384616 384617 384618 384621 384622 384623 384624 384625
#>   [51] 384626 384627 384630 384631 384633 384636 384637 384638 384639 384640
#>   [61] 384641 384642 384643 384644 384645 384950 384955 384956 384957 384960
#>   [71] 384963 384964 384965 384966 385079 385080 385081 385082 385085 385086
#>   [81] 385089 385090 385091 385092 385093 385094 385095 385097 385098 385101
#>   [91] 385102 385103 385104 385116 385118 385119 385120 385121 385122 385126
#>  [101] 385127 385128 385129 385130 385131 385132 385133 385134 385135 385137
#>  [111] 385138 385139 385140 385143 385144 385147 385149 385150 385153 385154
#>  [121] 385155 385156 385157 385158 385160 385161 385162 385163 385164 385168
#>  [131] 385173 385174 385175 385176 385181 385185 385186 385187 385188 385189
#>  [141] 385190 385191 385192 385193 385194 385196 385197 385201 385202 385204
#>  [151] 385205 385206 385207 385208 385209 385212 385214 385219 385220 385222
#>  [161] 385223 385226 385227 385232 385233 385234 385236 385240 385241 385242
#>  [171] 385244 385245 385248 385251 385253 385254 385259 385262 385263 385265
#>  [181] 385267 385268 385269 385270 385271 385274 385279 385280 385282 388039
#>  [191] 388041 388048 388049 388051 388052 388053 388054 388055 388056 388057
#>  [201] 388059 388061 388062 388063 388064 388065 388066 388067 388069 388070
#>  [211] 388071 388072 388073 388074 388075 388080 388081 388082 388083 388084
#>  [221] 388085 388086 388087 388088 388089 388092 388093 388094 388099 388100
#>  [231] 388101 388102 388103 388104 388105 388106 388109 388111 388112 388114
#>  [241] 388116 388117 388118 388119 388120 388123 388126 388127 388134 388135
#>  [251] 388136 388137 388138 388139 388140 388142 388144 388145 388147 388148
#>  [261] 388149 388150 388151 388152 388154 388156 388157 388158 388163 388164
#>  [271] 388165 388167 388168 388169 388170 388172 388173 388174 388175 388176
#>  [281] 388177 388178 388179 388181 388182 388183 388184 388185 388186 388188
#>  [291] 388189 388190 388192 388193 388194 388196 388197 388198 388199 388200
#>  [301] 388201 388202 388208 388209 388210 388211 388212 388214 388216 388217
#>  [311] 388218 388301 388302 388312 388336 388337 388338 388404 388407 388409
#>  [321] 388416 388417 388418 388419 388420 388421 388422 388423 388424 388425
#>  [331] 388427 388428 388429 388431 388432 388434 388435 388436 388437 388439
#>  [341] 388440 388444 388445 388446 388447 388448 388449 388450 388451 388452
#>  [351] 388454 388455 388456 388458 388459 388461 388463 388464 388465 388466
#>  [361] 388467 388469 388470 388472 388473 388474 388475 388476 388477 388479
#>  [371] 388480 388481 388482 388483 388485 388486 388490 388494 388495 388496
#>  [381] 388498 388499 388500 388502 388503 388504 388505 388512 388513 388514
#>  [391] 388515 388516 388519 388522 388523 388524 388527 388528 388529 388531
#>  [401] 388533 388534 388535 388536 388537 388538 388544 388545 388550 388551
#>  [411] 388552 388560 388561 388562 388564 388565 388566 388567 388568 388570
#>  [421] 388572 388574 388577 388578 388580 388582 388584 388585 388586 388587
#>  [431] 388588 388590 388591 388592 388593 388594 388597 388598 388599 388600
#>  [441] 388603 388604 388605 388606 388607 388608 388609 388612 388613 388614
#>  [451] 388616 388617 388618 388619 388620 388621 388623 388624 388625 388626
#>  [461] 388627 388628 388630 388634 388635 388636 388637 388638 388642 388643
#>  [471] 388646 388647 388650 388651 388652 388654 388655 388656 388658 388659
#>  [481] 388660 388661 388662 388663 388664 388666 388667 388668 388669 388670
#>  [491] 352688 352689 352690 352692 352693 352694 352691 352695 352696 352697
#>  [501] 352698 352699 352700 352701 352702 352703 352704 352705 352706 352707
#>  [511] 352708 352709 352710 352711 352712 352713 352716 352717 352718 352719
#>  [521] 352720 352721 352722 352723 352724 352725 352726 352727 353140 353141
#>  [531] 353142 353143 353145 353144 353146 353149 353150 353151 353152 353153
#>  [541] 353154 353155 353156 353157 353158 353159 353160 353161 353162 353163
#>  [551] 353164 353165 353167 353168 353169 353170 353171 353174 385100 385281
#>  [561] 388215 388426 388468 388579 388653 388665 395109 395110 395112 395113
#>  [571] 395115 395116 395120 395121 395122 395123 395124 395125 395126 395149
#>  [581] 395151 395152 395156 395170 395172 395173 395175 395177 395179 395181
#>  [591] 395182 395183 395184 395185 395186 395190 395191 384551 384552 384559
#>  [601] 384566 384567 384568 384571 384572 384573 384574 384578 384581 384583
#>  [611] 384585 384586 384588 384598 384619 384628 384629 384632 384634 384635
#>  [621] 384646 384647 384648 384951 384952 384953 384954 384958 384959 384961
#>  [631] 384962 385083 385084 385087 385088 385096 385099 385105 385106 385107
#>  [641] 385108 385109 385110 385111 385112 385113 385114 385115 385117 385123
#>  [651] 385124 385125 385136 385151 385159 385165 385166 385167 385169 385170
#>  [661] 385171 385172 385177 385178 385179 385180 385182 385183 385184 385195
#>  [671] 385198 385199 385200 385203 385210 385211 385213 385215 385216 385217
#>  [681] 385218 385224 385228 385231 385237 385238 385239 385243 385246 385249
#>  [691] 385250 385252 385256 385257 385258 385260 385261 385264 385266 385276
#>  [701] 388040 388042 388043 388044 388045 388046 388047 388050 388058 388060
#>  [711] 388076 388077 388078 388079 388090 388091 388095 388096 388097 388098
#>  [721] 388107 388108 388113 388115 388121 388122 388124 388125 388128 388129
#>  [731] 388130 388131 388132 388133 388141 388143 388146 388153 388155 388159
#>  [741] 388160 388161 388162 388171 388180 388187 388195 388213 388219 388220
#>  [751] 388221 388405 388406 388408 388410 388411 388412 388413 388414 388415
#>  [761] 388433 388438 388441 388442 388443 388453 388460 388471 388478 388487
#>  [771] 388488 388489 388491 388492 388493 388506 388507 388508 388509 388510
#>  [781] 388511 388517 388518 388521 388526 388530 388539 388540 388541 388542
#>  [791] 388547 388548 388549 388553 388555 388557 388559 388563 388569 388571
#>  [801] 388573 388575 388576 388581 388583 388601 388602 388611 388615 388622
#>  [811] 388629 388639 388640 388648 388649 388657 391944 395111 395117 395150
#>  [821] 395153 395154 395171 395174 395176 395180 395400 395401 352714 352715
#>  [831] 353136 353147 353148 353172 353173 385235 388191 388497 388631 395114
#>  [841] 395118 395119 395178 395192 410462 410463 410464 410465 410466 410467
#>  [851] 410469 410470 410472 410473 410474 410475 410476 410477 410478 410479
#>  [861] 410480 410481 410482 410483 410484 410485 410486 410489 410490 410491
#>  [871] 410492 410493 410495 410496 410497 410498 410499 410500 410501 410502
#>  [881] 410503 410504 410505 410507 410508 410509 410510 410511 410512 410513
#>  [891] 410516 410517 410518 410519 410520 410521 410522 410523 410524 410525
#>  [901] 410526 410527 410528 410529 410530 410531 410532 410533 410534 410535
#>  [911] 410536 410537 410539 410540 410541 410542 410544 410545 410546 410547
#>  [921] 410548 410549 410550 410551 410552 410555 410556 410558 410559 410561
#>  [931] 410563 410564 410565 410566 410567 410568 410569 410570 410571 410572
#>  [941] 410573 410574 410575 410576 410577 410578 410579 410580 410581 410582
#>  [951] 410583 410584 410585 410586 410587 410590 410591 410592 410593 410594
#>  [961] 410595 410596 410598 410599 410600 410601 410603 410604 410605 410606
#>  [971] 410607 410608 410609 410610 410611 410613 410614 410616 410619 410620
#>  [981] 410621 410622 410624 410625 410626 410627 410628 410629 410630 410631
#>  [991] 410632 410633 410634 410635 410637 410640 410641 410642 410643 410644
#> [1001] 410645 410646 410647 410648 410649 410650 410651 410652 410654 410655
#> [1011] 410656 410657 410658 410659 410660 410661 410662 410663 410664 410665
#> [1021] 410666 410667 410668 410669 410670 410671 410672 410673 410674 410675
#> [1031] 410676 410677 410678 410679 410680 410681 410682 410684 410685 410686
#> [1041] 410687 410688 410689 410690 410691 410692 410693 410694 410695 410696
#> [1051] 410697 410698 410699 410700 410701 410702 410703 410704 410705 410706
#> [1061] 410707 410708 410709 410710 410711 410712 410713 410714 410715 410716
#> [1071] 410717 410718 410719 410721 410722 410726 410727 410728 410730 410731
#> [1081] 410733 410734 410735 410736 410737 410738 410739 410740 410741 410742
#> [1091] 410743 410744 410745 410746 410747 410748 410749 410750 410751 410752
#> [1101] 410753 410754 410755 410756 410757 410758 410759 410760 410762 410763
#> [1111] 410764 410765 410766 410767 410768 410769 410770 410771 410772 410773
#> [1121] 410774 410775 410779 410782 410783 410784 410785 410786 410787 410788
#> [1131] 410791 410792 410793 410794 410795 410796 410797 410798 410799 410800
#> [1141] 410801 410803 410804 410805 410806 410807 410808 410809 410811 410812
#> [1151] 410813 410814 410815 410816 410818 410820 410821 410823 410825 410826
#> [1161] 410827 410828 410829 410830 410831 410832 410833 410834 410835 410836
#> [1171] 410837 410838 410839 410840 410841 410842 410843 410844 410845 410846
#> [1181] 410847 410848 410849 410850 410851 410852 410853 410855 410856 410857
#> [1191] 410858 410859 410860 410862 410863 410864 410865 410866 410867 410868
#> [1201] 410869 410870 410871 410872 410873 410874 410875 410876 410877 410878
#> [1211] 410879 410880 410881 410882 410883 410884 410885 410887 410888 410889
#> [1221] 410890 410891 410892 410894 410895 410897 410898 410899 410901 410906
#> [1231] 410907 410909 410912 410913 410914 410916 410917 410918 410919 410921
#> [1241] 410922 410923 410924 410925 410926 410927 410928 410930 410931 410932
#> [1251] 410933 410934 410935 410936 410937 410938 410942 410948 410949 410950
#> [1261] 410951 410952 410953 410954 410955 410956 410957 410958 410960 410961
#> [1271] 410962 410963 410964 410965 410966 410967 410968 410969 410970 410972
#> [1281] 410975 410976 410977 410978 410979 410980 410981 410982 410983 410984
#> [1291] 410985 410986 410987 410988 410989 410990 410991 410992 410993 410994
#> [1301] 410995 410996 410997 410998 410999 411000 411001 411681 411682 411683
#> [1311] 411685 411686 411687 411688 411689 411690 411691 411692 411693 411694
#> [1321] 411696 411697 411698 411699 411700 411701 411702 411703 411704 411705
#> [1331] 411706 411707 411708 411709 411710 411711 411712 411713 411714 411715
#> [1341] 411716 411717 411720 411721 411722 411724 411725 411730 411731 411732
#> [1351] 411733 411734 411735 411736 411738 411739 411740 411741 411742 411743
#> [1361] 411744 411745 411747 411748 411749 411752 411753 411758 411759 411761
#> [1371] 411762 411763 411764 411765 411766 411767 411768 411769 411770 411771
#> [1381] 411773 411774 411775 411776 411777 411780 411781 411782 411783 411784
#> [1391] 411785 411786 411787 411788 411790 411791 411792 411793 411794 411795
#> [1401] 411796 411797 411799 411801 411803 411805 411806 411807 411808 411809
#> [1411] 411810 411811 411813 411814 411815 411816 411818 411819 411822 411823
#> [1421] 411824 411825 411827 411828 411829 411832 411833 411834 411835 411836
#> [1431] 411837 411838 411839 411840 411841 411842 411843 411844 411846 411847
#> [1441] 411848 411849 411853 411855 411856 411857 412210 412211 412212 412213
#> [1451] 422362 422453 422483 423312 423313 423314 423334 423335 423336 423337
#> [1461] 423338 425236 425238 425244 425249 425254 425447 425692

p <- ggplot_tag_status_by_subquadrat(
  small_vft,
  title_quad = "BCI 2012. Quadrat: ",
  bl = "bottom-left", br = "bottom-right", tr = "top-right", tl = "top-left",
  header = "Line 1: _________\nLine 2:\nLine 3:.....................",
  subquad_offset = -1,
  point_size = 3, point_shape = c(17, 6),
  tag_size = 2,
  move_edge = 0.5
)

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

p[[1]]

#>    [1] 384550 384553 384554 384555 384556 384557 384558 384560 384561 384562
#>   [11] 384563 384564 384570 384575 384576 384577 384579 384580 384582 384584
#>   [21] 384587 384589 384590 384591 384592 384593 384594 384595 384596 384597
#>   [31] 384599 384600 384601 384602 384607 384608 384610 384611 384612 384613
#>   [41] 384614 384615 384616 384617 384618 384621 384622 384623 384624 384625
#>   [51] 384626 384627 384630 384631 384633 384636 384637 384638 384639 384640
#>   [61] 384641 384642 384643 384644 384645 384950 384955 384956 384957 384960
#>   [71] 384963 384964 384965 384966 385079 385080 385081 385082 385085 385086
#>   [81] 385089 385090 385091 385092 385093 385094 385095 385097 385098 385101
#>   [91] 385102 385103 385104 385116 385118 385119 385120 385121 385122 385126
#>  [101] 385127 385128 385129 385130 385131 385132 385133 385134 385135 385137
#>  [111] 385138 385139 385140 385143 385144 385147 385149 385150 385153 385154
#>  [121] 385155 385156 385157 385158 385160 385161 385162 385163 385164 385168
#>  [131] 385173 385174 385175 385176 385181 385185 385186 385187 385188 385189
#>  [141] 385190 385191 385192 385193 385194 385196 385197 385201 385202 385204
#>  [151] 385205 385206 385207 385208 385209 385212 385214 385219 385220 385222
#>  [161] 385223 385226 385227 385232 385233 385234 385236 385240 385241 385242
#>  [171] 385244 385245 385248 385251 385253 385254 385259 385262 385263 385265
#>  [181] 385267 385268 385269 385270 385271 385274 385279 385280 385282 388039
#>  [191] 388041 388048 388049 388051 388052 388053 388054 388055 388056 388057
#>  [201] 388059 388061 388062 388063 388064 388065 388066 388067 388069 388070
#>  [211] 388071 388072 388073 388074 388075 388080 388081 388082 388083 388084
#>  [221] 388085 388086 388087 388088 388089 388092 388093 388094 388099 388100
#>  [231] 388101 388102 388103 388104 388105 388106 388109 388111 388112 388114
#>  [241] 388116 388117 388118 388119 388120 388123 388126 388127 388134 388135
#>  [251] 388136 388137 388138 388139 388140 388142 388144 388145 388147 388148
#>  [261] 388149 388150 388151 388152 388154 388156 388157 388158 388163 388164
#>  [271] 388165 388167 388168 388169 388170 388172 388173 388174 388175 388176
#>  [281] 388177 388178 388179 388181 388182 388183 388184 388185 388186 388188
#>  [291] 388189 388190 388192 388193 388194 388196 388197 388198 388199 388200
#>  [301] 388201 388202 388208 388209 388210 388211 388212 388214 388216 388217
#>  [311] 388218 388301 388302 388312 388336 388337 388338 388404 388407 388409
#>  [321] 388416 388417 388418 388419 388420 388421 388422 388423 388424 388425
#>  [331] 388427 388428 388429 388431 388432 388434 388435 388436 388437 388439
#>  [341] 388440 388444 388445 388446 388447 388448 388449 388450 388451 388452
#>  [351] 388454 388455 388456 388458 388459 388461 388463 388464 388465 388466
#>  [361] 388467 388469 388470 388472 388473 388474 388475 388476 388477 388479
#>  [371] 388480 388481 388482 388483 388485 388486 388490 388494 388495 388496
#>  [381] 388498 388499 388500 388502 388503 388504 388505 388512 388513 388514
#>  [391] 388515 388516 388519 388522 388523 388524 388527 388528 388529 388531
#>  [401] 388533 388534 388535 388536 388537 388538 388544 388545 388550 388551
#>  [411] 388552 388560 388561 388562 388564 388565 388566 388567 388568 388570
#>  [421] 388572 388574 388577 388578 388580 388582 388584 388585 388586 388587
#>  [431] 388588 388590 388591 388592 388593 388594 388597 388598 388599 388600
#>  [441] 388603 388604 388605 388606 388607 388608 388609 388612 388613 388614
#>  [451] 388616 388617 388618 388619 388620 388621 388623 388624 388625 388626
#>  [461] 388627 388628 388630 388634 388635 388636 388637 388638 388642 388643
#>  [471] 388646 388647 388650 388651 388652 388654 388655 388656 388658 388659
#>  [481] 388660 388661 388662 388663 388664 388666 388667 388668 388669 388670
#>  [491] 352688 352689 352690 352692 352693 352694 352691 352695 352696 352697
#>  [501] 352698 352699 352700 352701 352702 352703 352704 352705 352706 352707
#>  [511] 352708 352709 352710 352711 352712 352713 352716 352717 352718 352719
#>  [521] 352720 352721 352722 352723 352724 352725 352726 352727 353140 353141
#>  [531] 353142 353143 353145 353144 353146 353149 353150 353151 353152 353153
#>  [541] 353154 353155 353156 353157 353158 353159 353160 353161 353162 353163
#>  [551] 353164 353165 353167 353168 353169 353170 353171 353174 385100 385281
#>  [561] 388215 388426 388468 388579 388653 388665 395109 395110 395112 395113
#>  [571] 395115 395116 395120 395121 395122 395123 395124 395125 395126 395149
#>  [581] 395151 395152 395156 395170 395172 395173 395175 395177 395179 395181
#>  [591] 395182 395183 395184 395185 395186 395190 395191 384551 384552 384559
#>  [601] 384566 384567 384568 384571 384572 384573 384574 384578 384581 384583
#>  [611] 384585 384586 384588 384598 384619 384628 384629 384632 384634 384635
#>  [621] 384646 384647 384648 384951 384952 384953 384954 384958 384959 384961
#>  [631] 384962 385083 385084 385087 385088 385096 385099 385105 385106 385107
#>  [641] 385108 385109 385110 385111 385112 385113 385114 385115 385117 385123
#>  [651] 385124 385125 385136 385151 385159 385165 385166 385167 385169 385170
#>  [661] 385171 385172 385177 385178 385179 385180 385182 385183 385184 385195
#>  [671] 385198 385199 385200 385203 385210 385211 385213 385215 385216 385217
#>  [681] 385218 385224 385228 385231 385237 385238 385239 385243 385246 385249
#>  [691] 385250 385252 385256 385257 385258 385260 385261 385264 385266 385276
#>  [701] 388040 388042 388043 388044 388045 388046 388047 388050 388058 388060
#>  [711] 388076 388077 388078 388079 388090 388091 388095 388096 388097 388098
#>  [721] 388107 388108 388113 388115 388121 388122 388124 388125 388128 388129
#>  [731] 388130 388131 388132 388133 388141 388143 388146 388153 388155 388159
#>  [741] 388160 388161 388162 388171 388180 388187 388195 388213 388219 388220
#>  [751] 388221 388405 388406 388408 388410 388411 388412 388413 388414 388415
#>  [761] 388433 388438 388441 388442 388443 388453 388460 388471 388478 388487
#>  [771] 388488 388489 388491 388492 388493 388506 388507 388508 388509 388510
#>  [781] 388511 388517 388518 388521 388526 388530 388539 388540 388541 388542
#>  [791] 388547 388548 388549 388553 388555 388557 388559 388563 388569 388571
#>  [801] 388573 388575 388576 388581 388583 388601 388602 388611 388615 388622
#>  [811] 388629 388639 388640 388648 388649 388657 391944 395111 395117 395150
#>  [821] 395153 395154 395171 395174 395176 395180 395400 395401 352714 352715
#>  [831] 353136 353147 353148 353172 353173 385235 388191 388497 388631 395114
#>  [841] 395118 395119 395178 395192 410462 410463 410464 410465 410466 410467
#>  [851] 410469 410470 410472 410473 410474 410475 410476 410477 410478 410479
#>  [861] 410480 410481 410482 410483 410484 410485 410486 410489 410490 410491
#>  [871] 410492 410493 410495 410496 410497 410498 410499 410500 410501 410502
#>  [881] 410503 410504 410505 410507 410508 410509 410510 410511 410512 410513
#>  [891] 410516 410517 410518 410519 410520 410521 410522 410523 410524 410525
#>  [901] 410526 410527 410528 410529 410530 410531 410532 410533 410534 410535
#>  [911] 410536 410537 410539 410540 410541 410542 410544 410545 410546 410547
#>  [921] 410548 410549 410550 410551 410552 410555 410556 410558 410559 410561
#>  [931] 410563 410564 410565 410566 410567 410568 410569 410570 410571 410572
#>  [941] 410573 410574 410575 410576 410577 410578 410579 410580 410581 410582
#>  [951] 410583 410584 410585 410586 410587 410590 410591 410592 410593 410594
#>  [961] 410595 410596 410598 410599 410600 410601 410603 410604 410605 410606
#>  [971] 410607 410608 410609 410610 410611 410613 410614 410616 410619 410620
#>  [981] 410621 410622 410624 410625 410626 410627 410628 410629 410630 410631
#>  [991] 410632 410633 410634 410635 410637 410640 410641 410642 410643 410644
#> [1001] 410645 410646 410647 410648 410649 410650 410651 410652 410654 410655
#> [1011] 410656 410657 410658 410659 410660 410661 410662 410663 410664 410665
#> [1021] 410666 410667 410668 410669 410670 410671 410672 410673 410674 410675
#> [1031] 410676 410677 410678 410679 410680 410681 410682 410684 410685 410686
#> [1041] 410687 410688 410689 410690 410691 410692 410693 410694 410695 410696
#> [1051] 410697 410698 410699 410700 410701 410702 410703 410704 410705 410706
#> [1061] 410707 410708 410709 410710 410711 410712 410713 410714 410715 410716
#> [1071] 410717 410718 410719 410721 410722 410726 410727 410728 410730 410731
#> [1081] 410733 410734 410735 410736 410737 410738 410739 410740 410741 410742
#> [1091] 410743 410744 410745 410746 410747 410748 410749 410750 410751 410752
#> [1101] 410753 410754 410755 410756 410757 410758 410759 410760 410762 410763
#> [1111] 410764 410765 410766 410767 410768 410769 410770 410771 410772 410773
#> [1121] 410774 410775 410779 410782 410783 410784 410785 410786 410787 410788
#> [1131] 410791 410792 410793 410794 410795 410796 410797 410798 410799 410800
#> [1141] 410801 410803 410804 410805 410806 410807 410808 410809 410811 410812
#> [1151] 410813 410814 410815 410816 410818 410820 410821 410823 410825 410826
#> [1161] 410827 410828 410829 410830 410831 410832 410833 410834 410835 410836
#> [1171] 410837 410838 410839 410840 410841 410842 410843 410844 410845 410846
#> [1181] 410847 410848 410849 410850 410851 410852 410853 410855 410856 410857
#> [1191] 410858 410859 410860 410862 410863 410864 410865 410866 410867 410868
#> [1201] 410869 410870 410871 410872 410873 410874 410875 410876 410877 410878
#> [1211] 410879 410880 410881 410882 410883 410884 410885 410887 410888 410889
#> [1221] 410890 410891 410892 410894 410895 410897 410898 410899 410901 410906
#> [1231] 410907 410909 410912 410913 410914 410916 410917 410918 410919 410921
#> [1241] 410922 410923 410924 410925 410926 410927 410928 410930 410931 410932
#> [1251] 410933 410934 410935 410936 410937 410938 410942 410948 410949 410950
#> [1261] 410951 410952 410953 410954 410955 410956 410957 410958 410960 410961
#> [1271] 410962 410963 410964 410965 410966 410967 410968 410969 410970 410972
#> [1281] 410975 410976 410977 410978 410979 410980 410981 410982 410983 410984
#> [1291] 410985 410986 410987 410988 410989 410990 410991 410992 410993 410994
#> [1301] 410995 410996 410997 410998 410999 411000 411001 411681 411682 411683
#> [1311] 411685 411686 411687 411688 411689 411690 411691 411692 411693 411694
#> [1321] 411696 411697 411698 411699 411700 411701 411702 411703 411704 411705
#> [1331] 411706 411707 411708 411709 411710 411711 411712 411713 411714 411715
#> [1341] 411716 411717 411720 411721 411722 411724 411725 411730 411731 411732
#> [1351] 411733 411734 411735 411736 411738 411739 411740 411741 411742 411743
#> [1361] 411744 411745 411747 411748 411749 411752 411753 411758 411759 411761
#> [1371] 411762 411763 411764 411765 411766 411767 411768 411769 411770 411771
#> [1381] 411773 411774 411775 411776 411777 411780 411781 411782 411783 411784
#> [1391] 411785 411786 411787 411788 411790 411791 411792 411793 411794 411795
#> [1401] 411796 411797 411799 411801 411803 411805 411806 411807 411808 411809
#> [1411] 411810 411811 411813 411814 411815 411816 411818 411819 411822 411823
#> [1421] 411824 411825 411827 411828 411829 411832 411833 411834 411835 411836
#> [1431] 411837 411838 411839 411840 411841 411842 411843 411844 411846 411847
#> [1441] 411848 411849 411853 411855 411856 411857 412210 412211 412212 412213
#> [1451] 422362 422453 422483 423312 423313 423314 423334 423335 423336 423337
#> [1461] 423338 425236 425238 425244 425249 425254 425447 425692

# Themes
library(ggplot2)

p <- ggplot_tag_status_by_subquadrat(small_vft, theme = theme_gray())

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

p[[1]]

#>    [1] 384550 384553 384554 384555 384556 384557 384558 384560 384561 384562
#>   [11] 384563 384564 384570 384575 384576 384577 384579 384580 384582 384584
#>   [21] 384587 384589 384590 384591 384592 384593 384594 384595 384596 384597
#>   [31] 384599 384600 384601 384602 384607 384608 384610 384611 384612 384613
#>   [41] 384614 384615 384616 384617 384618 384621 384622 384623 384624 384625
#>   [51] 384626 384627 384630 384631 384633 384636 384637 384638 384639 384640
#>   [61] 384641 384642 384643 384644 384645 384950 384955 384956 384957 384960
#>   [71] 384963 384964 384965 384966 385079 385080 385081 385082 385085 385086
#>   [81] 385089 385090 385091 385092 385093 385094 385095 385097 385098 385101
#>   [91] 385102 385103 385104 385116 385118 385119 385120 385121 385122 385126
#>  [101] 385127 385128 385129 385130 385131 385132 385133 385134 385135 385137
#>  [111] 385138 385139 385140 385143 385144 385147 385149 385150 385153 385154
#>  [121] 385155 385156 385157 385158 385160 385161 385162 385163 385164 385168
#>  [131] 385173 385174 385175 385176 385181 385185 385186 385187 385188 385189
#>  [141] 385190 385191 385192 385193 385194 385196 385197 385201 385202 385204
#>  [151] 385205 385206 385207 385208 385209 385212 385214 385219 385220 385222
#>  [161] 385223 385226 385227 385232 385233 385234 385236 385240 385241 385242
#>  [171] 385244 385245 385248 385251 385253 385254 385259 385262 385263 385265
#>  [181] 385267 385268 385269 385270 385271 385274 385279 385280 385282 388039
#>  [191] 388041 388048 388049 388051 388052 388053 388054 388055 388056 388057
#>  [201] 388059 388061 388062 388063 388064 388065 388066 388067 388069 388070
#>  [211] 388071 388072 388073 388074 388075 388080 388081 388082 388083 388084
#>  [221] 388085 388086 388087 388088 388089 388092 388093 388094 388099 388100
#>  [231] 388101 388102 388103 388104 388105 388106 388109 388111 388112 388114
#>  [241] 388116 388117 388118 388119 388120 388123 388126 388127 388134 388135
#>  [251] 388136 388137 388138 388139 388140 388142 388144 388145 388147 388148
#>  [261] 388149 388150 388151 388152 388154 388156 388157 388158 388163 388164
#>  [271] 388165 388167 388168 388169 388170 388172 388173 388174 388175 388176
#>  [281] 388177 388178 388179 388181 388182 388183 388184 388185 388186 388188
#>  [291] 388189 388190 388192 388193 388194 388196 388197 388198 388199 388200
#>  [301] 388201 388202 388208 388209 388210 388211 388212 388214 388216 388217
#>  [311] 388218 388301 388302 388312 388336 388337 388338 388404 388407 388409
#>  [321] 388416 388417 388418 388419 388420 388421 388422 388423 388424 388425
#>  [331] 388427 388428 388429 388431 388432 388434 388435 388436 388437 388439
#>  [341] 388440 388444 388445 388446 388447 388448 388449 388450 388451 388452
#>  [351] 388454 388455 388456 388458 388459 388461 388463 388464 388465 388466
#>  [361] 388467 388469 388470 388472 388473 388474 388475 388476 388477 388479
#>  [371] 388480 388481 388482 388483 388485 388486 388490 388494 388495 388496
#>  [381] 388498 388499 388500 388502 388503 388504 388505 388512 388513 388514
#>  [391] 388515 388516 388519 388522 388523 388524 388527 388528 388529 388531
#>  [401] 388533 388534 388535 388536 388537 388538 388544 388545 388550 388551
#>  [411] 388552 388560 388561 388562 388564 388565 388566 388567 388568 388570
#>  [421] 388572 388574 388577 388578 388580 388582 388584 388585 388586 388587
#>  [431] 388588 388590 388591 388592 388593 388594 388597 388598 388599 388600
#>  [441] 388603 388604 388605 388606 388607 388608 388609 388612 388613 388614
#>  [451] 388616 388617 388618 388619 388620 388621 388623 388624 388625 388626
#>  [461] 388627 388628 388630 388634 388635 388636 388637 388638 388642 388643
#>  [471] 388646 388647 388650 388651 388652 388654 388655 388656 388658 388659
#>  [481] 388660 388661 388662 388663 388664 388666 388667 388668 388669 388670
#>  [491] 352688 352689 352690 352692 352693 352694 352691 352695 352696 352697
#>  [501] 352698 352699 352700 352701 352702 352703 352704 352705 352706 352707
#>  [511] 352708 352709 352710 352711 352712 352713 352716 352717 352718 352719
#>  [521] 352720 352721 352722 352723 352724 352725 352726 352727 353140 353141
#>  [531] 353142 353143 353145 353144 353146 353149 353150 353151 353152 353153
#>  [541] 353154 353155 353156 353157 353158 353159 353160 353161 353162 353163
#>  [551] 353164 353165 353167 353168 353169 353170 353171 353174 385100 385281
#>  [561] 388215 388426 388468 388579 388653 388665 395109 395110 395112 395113
#>  [571] 395115 395116 395120 395121 395122 395123 395124 395125 395126 395149
#>  [581] 395151 395152 395156 395170 395172 395173 395175 395177 395179 395181
#>  [591] 395182 395183 395184 395185 395186 395190 395191 384551 384552 384559
#>  [601] 384566 384567 384568 384571 384572 384573 384574 384578 384581 384583
#>  [611] 384585 384586 384588 384598 384619 384628 384629 384632 384634 384635
#>  [621] 384646 384647 384648 384951 384952 384953 384954 384958 384959 384961
#>  [631] 384962 385083 385084 385087 385088 385096 385099 385105 385106 385107
#>  [641] 385108 385109 385110 385111 385112 385113 385114 385115 385117 385123
#>  [651] 385124 385125 385136 385151 385159 385165 385166 385167 385169 385170
#>  [661] 385171 385172 385177 385178 385179 385180 385182 385183 385184 385195
#>  [671] 385198 385199 385200 385203 385210 385211 385213 385215 385216 385217
#>  [681] 385218 385224 385228 385231 385237 385238 385239 385243 385246 385249
#>  [691] 385250 385252 385256 385257 385258 385260 385261 385264 385266 385276
#>  [701] 388040 388042 388043 388044 388045 388046 388047 388050 388058 388060
#>  [711] 388076 388077 388078 388079 388090 388091 388095 388096 388097 388098
#>  [721] 388107 388108 388113 388115 388121 388122 388124 388125 388128 388129
#>  [731] 388130 388131 388132 388133 388141 388143 388146 388153 388155 388159
#>  [741] 388160 388161 388162 388171 388180 388187 388195 388213 388219 388220
#>  [751] 388221 388405 388406 388408 388410 388411 388412 388413 388414 388415
#>  [761] 388433 388438 388441 388442 388443 388453 388460 388471 388478 388487
#>  [771] 388488 388489 388491 388492 388493 388506 388507 388508 388509 388510
#>  [781] 388511 388517 388518 388521 388526 388530 388539 388540 388541 388542
#>  [791] 388547 388548 388549 388553 388555 388557 388559 388563 388569 388571
#>  [801] 388573 388575 388576 388581 388583 388601 388602 388611 388615 388622
#>  [811] 388629 388639 388640 388648 388649 388657 391944 395111 395117 395150
#>  [821] 395153 395154 395171 395174 395176 395180 395400 395401 352714 352715
#>  [831] 353136 353147 353148 353172 353173 385235 388191 388497 388631 395114
#>  [841] 395118 395119 395178 395192 410462 410463 410464 410465 410466 410467
#>  [851] 410469 410470 410472 410473 410474 410475 410476 410477 410478 410479
#>  [861] 410480 410481 410482 410483 410484 410485 410486 410489 410490 410491
#>  [871] 410492 410493 410495 410496 410497 410498 410499 410500 410501 410502
#>  [881] 410503 410504 410505 410507 410508 410509 410510 410511 410512 410513
#>  [891] 410516 410517 410518 410519 410520 410521 410522 410523 410524 410525
#>  [901] 410526 410527 410528 410529 410530 410531 410532 410533 410534 410535
#>  [911] 410536 410537 410539 410540 410541 410542 410544 410545 410546 410547
#>  [921] 410548 410549 410550 410551 410552 410555 410556 410558 410559 410561
#>  [931] 410563 410564 410565 410566 410567 410568 410569 410570 410571 410572
#>  [941] 410573 410574 410575 410576 410577 410578 410579 410580 410581 410582
#>  [951] 410583 410584 410585 410586 410587 410590 410591 410592 410593 410594
#>  [961] 410595 410596 410598 410599 410600 410601 410603 410604 410605 410606
#>  [971] 410607 410608 410609 410610 410611 410613 410614 410616 410619 410620
#>  [981] 410621 410622 410624 410625 410626 410627 410628 410629 410630 410631
#>  [991] 410632 410633 410634 410635 410637 410640 410641 410642 410643 410644
#> [1001] 410645 410646 410647 410648 410649 410650 410651 410652 410654 410655
#> [1011] 410656 410657 410658 410659 410660 410661 410662 410663 410664 410665
#> [1021] 410666 410667 410668 410669 410670 410671 410672 410673 410674 410675
#> [1031] 410676 410677 410678 410679 410680 410681 410682 410684 410685 410686
#> [1041] 410687 410688 410689 410690 410691 410692 410693 410694 410695 410696
#> [1051] 410697 410698 410699 410700 410701 410702 410703 410704 410705 410706
#> [1061] 410707 410708 410709 410710 410711 410712 410713 410714 410715 410716
#> [1071] 410717 410718 410719 410721 410722 410726 410727 410728 410730 410731
#> [1081] 410733 410734 410735 410736 410737 410738 410739 410740 410741 410742
#> [1091] 410743 410744 410745 410746 410747 410748 410749 410750 410751 410752
#> [1101] 410753 410754 410755 410756 410757 410758 410759 410760 410762 410763
#> [1111] 410764 410765 410766 410767 410768 410769 410770 410771 410772 410773
#> [1121] 410774 410775 410779 410782 410783 410784 410785 410786 410787 410788
#> [1131] 410791 410792 410793 410794 410795 410796 410797 410798 410799 410800
#> [1141] 410801 410803 410804 410805 410806 410807 410808 410809 410811 410812
#> [1151] 410813 410814 410815 410816 410818 410820 410821 410823 410825 410826
#> [1161] 410827 410828 410829 410830 410831 410832 410833 410834 410835 410836
#> [1171] 410837 410838 410839 410840 410841 410842 410843 410844 410845 410846
#> [1181] 410847 410848 410849 410850 410851 410852 410853 410855 410856 410857
#> [1191] 410858 410859 410860 410862 410863 410864 410865 410866 410867 410868
#> [1201] 410869 410870 410871 410872 410873 410874 410875 410876 410877 410878
#> [1211] 410879 410880 410881 410882 410883 410884 410885 410887 410888 410889
#> [1221] 410890 410891 410892 410894 410895 410897 410898 410899 410901 410906
#> [1231] 410907 410909 410912 410913 410914 410916 410917 410918 410919 410921
#> [1241] 410922 410923 410924 410925 410926 410927 410928 410930 410931 410932
#> [1251] 410933 410934 410935 410936 410937 410938 410942 410948 410949 410950
#> [1261] 410951 410952 410953 410954 410955 410956 410957 410958 410960 410961
#> [1271] 410962 410963 410964 410965 410966 410967 410968 410969 410970 410972
#> [1281] 410975 410976 410977 410978 410979 410980 410981 410982 410983 410984
#> [1291] 410985 410986 410987 410988 410989 410990 410991 410992 410993 410994
#> [1301] 410995 410996 410997 410998 410999 411000 411001 411681 411682 411683
#> [1311] 411685 411686 411687 411688 411689 411690 411691 411692 411693 411694
#> [1321] 411696 411697 411698 411699 411700 411701 411702 411703 411704 411705
#> [1331] 411706 411707 411708 411709 411710 411711 411712 411713 411714 411715
#> [1341] 411716 411717 411720 411721 411722 411724 411725 411730 411731 411732
#> [1351] 411733 411734 411735 411736 411738 411739 411740 411741 411742 411743
#> [1361] 411744 411745 411747 411748 411749 411752 411753 411758 411759 411761
#> [1371] 411762 411763 411764 411765 411766 411767 411768 411769 411770 411771
#> [1381] 411773 411774 411775 411776 411777 411780 411781 411782 411783 411784
#> [1391] 411785 411786 411787 411788 411790 411791 411792 411793 411794 411795
#> [1401] 411796 411797 411799 411801 411803 411805 411806 411807 411808 411809
#> [1411] 411810 411811 411813 411814 411815 411816 411818 411819 411822 411823
#> [1421] 411824 411825 411827 411828 411829 411832 411833 411834 411835 411836
#> [1431] 411837 411838 411839 411840 411841 411842 411843 411844 411846 411847
#> [1441] 411848 411849 411853 411855 411856 411857 412210 412211 412212 412213
#> [1451] 422362 422453 422483 423312 423313 423314 423334 423335 423336 423337
#> [1461] 423338 425236 425238 425244 425249 425254 425447 425692

# Tweaking the default theme of ggplot_tag_status_by_subquadrat()

small_tweak <- theme_tag_status(legend.position = "bottom")
p <- ggplot_tag_status_by_subquadrat(small_vft, theme = small_tweak)

#> Error in ggplot(data = prep_df, aes(x = qx, y = qy, shape = status_tree)): unused argument (aes(x = qx, y = qy, shape = status_tree))

p[[1]]

#>    [1] 384550 384553 384554 384555 384556 384557 384558 384560 384561 384562
#>   [11] 384563 384564 384570 384575 384576 384577 384579 384580 384582 384584
#>   [21] 384587 384589 384590 384591 384592 384593 384594 384595 384596 384597
#>   [31] 384599 384600 384601 384602 384607 384608 384610 384611 384612 384613
#>   [41] 384614 384615 384616 384617 384618 384621 384622 384623 384624 384625
#>   [51] 384626 384627 384630 384631 384633 384636 384637 384638 384639 384640
#>   [61] 384641 384642 384643 384644 384645 384950 384955 384956 384957 384960
#>   [71] 384963 384964 384965 384966 385079 385080 385081 385082 385085 385086
#>   [81] 385089 385090 385091 385092 385093 385094 385095 385097 385098 385101
#>   [91] 385102 385103 385104 385116 385118 385119 385120 385121 385122 385126
#>  [101] 385127 385128 385129 385130 385131 385132 385133 385134 385135 385137
#>  [111] 385138 385139 385140 385143 385144 385147 385149 385150 385153 385154
#>  [121] 385155 385156 385157 385158 385160 385161 385162 385163 385164 385168
#>  [131] 385173 385174 385175 385176 385181 385185 385186 385187 385188 385189
#>  [141] 385190 385191 385192 385193 385194 385196 385197 385201 385202 385204
#>  [151] 385205 385206 385207 385208 385209 385212 385214 385219 385220 385222
#>  [161] 385223 385226 385227 385232 385233 385234 385236 385240 385241 385242
#>  [171] 385244 385245 385248 385251 385253 385254 385259 385262 385263 385265
#>  [181] 385267 385268 385269 385270 385271 385274 385279 385280 385282 388039
#>  [191] 388041 388048 388049 388051 388052 388053 388054 388055 388056 388057
#>  [201] 388059 388061 388062 388063 388064 388065 388066 388067 388069 388070
#>  [211] 388071 388072 388073 388074 388075 388080 388081 388082 388083 388084
#>  [221] 388085 388086 388087 388088 388089 388092 388093 388094 388099 388100
#>  [231] 388101 388102 388103 388104 388105 388106 388109 388111 388112 388114
#>  [241] 388116 388117 388118 388119 388120 388123 388126 388127 388134 388135
#>  [251] 388136 388137 388138 388139 388140 388142 388144 388145 388147 388148
#>  [261] 388149 388150 388151 388152 388154 388156 388157 388158 388163 388164
#>  [271] 388165 388167 388168 388169 388170 388172 388173 388174 388175 388176
#>  [281] 388177 388178 388179 388181 388182 388183 388184 388185 388186 388188
#>  [291] 388189 388190 388192 388193 388194 388196 388197 388198 388199 388200
#>  [301] 388201 388202 388208 388209 388210 388211 388212 388214 388216 388217
#>  [311] 388218 388301 388302 388312 388336 388337 388338 388404 388407 388409
#>  [321] 388416 388417 388418 388419 388420 388421 388422 388423 388424 388425
#>  [331] 388427 388428 388429 388431 388432 388434 388435 388436 388437 388439
#>  [341] 388440 388444 388445 388446 388447 388448 388449 388450 388451 388452
#>  [351] 388454 388455 388456 388458 388459 388461 388463 388464 388465 388466
#>  [361] 388467 388469 388470 388472 388473 388474 388475 388476 388477 388479
#>  [371] 388480 388481 388482 388483 388485 388486 388490 388494 388495 388496
#>  [381] 388498 388499 388500 388502 388503 388504 388505 388512 388513 388514
#>  [391] 388515 388516 388519 388522 388523 388524 388527 388528 388529 388531
#>  [401] 388533 388534 388535 388536 388537 388538 388544 388545 388550 388551
#>  [411] 388552 388560 388561 388562 388564 388565 388566 388567 388568 388570
#>  [421] 388572 388574 388577 388578 388580 388582 388584 388585 388586 388587
#>  [431] 388588 388590 388591 388592 388593 388594 388597 388598 388599 388600
#>  [441] 388603 388604 388605 388606 388607 388608 388609 388612 388613 388614
#>  [451] 388616 388617 388618 388619 388620 388621 388623 388624 388625 388626
#>  [461] 388627 388628 388630 388634 388635 388636 388637 388638 388642 388643
#>  [471] 388646 388647 388650 388651 388652 388654 388655 388656 388658 388659
#>  [481] 388660 388661 388662 388663 388664 388666 388667 388668 388669 388670
#>  [491] 352688 352689 352690 352692 352693 352694 352691 352695 352696 352697
#>  [501] 352698 352699 352700 352701 352702 352703 352704 352705 352706 352707
#>  [511] 352708 352709 352710 352711 352712 352713 352716 352717 352718 352719
#>  [521] 352720 352721 352722 352723 352724 352725 352726 352727 353140 353141
#>  [531] 353142 353143 353145 353144 353146 353149 353150 353151 353152 353153
#>  [541] 353154 353155 353156 353157 353158 353159 353160 353161 353162 353163
#>  [551] 353164 353165 353167 353168 353169 353170 353171 353174 385100 385281
#>  [561] 388215 388426 388468 388579 388653 388665 395109 395110 395112 395113
#>  [571] 395115 395116 395120 395121 395122 395123 395124 395125 395126 395149
#>  [581] 395151 395152 395156 395170 395172 395173 395175 395177 395179 395181
#>  [591] 395182 395183 395184 395185 395186 395190 395191 384551 384552 384559
#>  [601] 384566 384567 384568 384571 384572 384573 384574 384578 384581 384583
#>  [611] 384585 384586 384588 384598 384619 384628 384629 384632 384634 384635
#>  [621] 384646 384647 384648 384951 384952 384953 384954 384958 384959 384961
#>  [631] 384962 385083 385084 385087 385088 385096 385099 385105 385106 385107
#>  [641] 385108 385109 385110 385111 385112 385113 385114 385115 385117 385123
#>  [651] 385124 385125 385136 385151 385159 385165 385166 385167 385169 385170
#>  [661] 385171 385172 385177 385178 385179 385180 385182 385183 385184 385195
#>  [671] 385198 385199 385200 385203 385210 385211 385213 385215 385216 385217
#>  [681] 385218 385224 385228 385231 385237 385238 385239 385243 385246 385249
#>  [691] 385250 385252 385256 385257 385258 385260 385261 385264 385266 385276
#>  [701] 388040 388042 388043 388044 388045 388046 388047 388050 388058 388060
#>  [711] 388076 388077 388078 388079 388090 388091 388095 388096 388097 388098
#>  [721] 388107 388108 388113 388115 388121 388122 388124 388125 388128 388129
#>  [731] 388130 388131 388132 388133 388141 388143 388146 388153 388155 388159
#>  [741] 388160 388161 388162 388171 388180 388187 388195 388213 388219 388220
#>  [751] 388221 388405 388406 388408 388410 388411 388412 388413 388414 388415
#>  [761] 388433 388438 388441 388442 388443 388453 388460 388471 388478 388487
#>  [771] 388488 388489 388491 388492 388493 388506 388507 388508 388509 388510
#>  [781] 388511 388517 388518 388521 388526 388530 388539 388540 388541 388542
#>  [791] 388547 388548 388549 388553 388555 388557 388559 388563 388569 388571
#>  [801] 388573 388575 388576 388581 388583 388601 388602 388611 388615 388622
#>  [811] 388629 388639 388640 388648 388649 388657 391944 395111 395117 395150
#>  [821] 395153 395154 395171 395174 395176 395180 395400 395401 352714 352715
#>  [831] 353136 353147 353148 353172 353173 385235 388191 388497 388631 395114
#>  [841] 395118 395119 395178 395192 410462 410463 410464 410465 410466 410467
#>  [851] 410469 410470 410472 410473 410474 410475 410476 410477 410478 410479
#>  [861] 410480 410481 410482 410483 410484 410485 410486 410489 410490 410491
#>  [871] 410492 410493 410495 410496 410497 410498 410499 410500 410501 410502
#>  [881] 410503 410504 410505 410507 410508 410509 410510 410511 410512 410513
#>  [891] 410516 410517 410518 410519 410520 410521 410522 410523 410524 410525
#>  [901] 410526 410527 410528 410529 410530 410531 410532 410533 410534 410535
#>  [911] 410536 410537 410539 410540 410541 410542 410544 410545 410546 410547
#>  [921] 410548 410549 410550 410551 410552 410555 410556 410558 410559 410561
#>  [931] 410563 410564 410565 410566 410567 410568 410569 410570 410571 410572
#>  [941] 410573 410574 410575 410576 410577 410578 410579 410580 410581 410582
#>  [951] 410583 410584 410585 410586 410587 410590 410591 410592 410593 410594
#>  [961] 410595 410596 410598 410599 410600 410601 410603 410604 410605 410606
#>  [971] 410607 410608 410609 410610 410611 410613 410614 410616 410619 410620
#>  [981] 410621 410622 410624 410625 410626 410627 410628 410629 410630 410631
#>  [991] 410632 410633 410634 410635 410637 410640 410641 410642 410643 410644
#> [1001] 410645 410646 410647 410648 410649 410650 410651 410652 410654 410655
#> [1011] 410656 410657 410658 410659 410660 410661 410662 410663 410664 410665
#> [1021] 410666 410667 410668 410669 410670 410671 410672 410673 410674 410675
#> [1031] 410676 410677 410678 410679 410680 410681 410682 410684 410685 410686
#> [1041] 410687 410688 410689 410690 410691 410692 410693 410694 410695 410696
#> [1051] 410697 410698 410699 410700 410701 410702 410703 410704 410705 410706
#> [1061] 410707 410708 410709 410710 410711 410712 410713 410714 410715 410716
#> [1071] 410717 410718 410719 410721 410722 410726 410727 410728 410730 410731
#> [1081] 410733 410734 410735 410736 410737 410738 410739 410740 410741 410742
#> [1091] 410743 410744 410745 410746 410747 410748 410749 410750 410751 410752
#> [1101] 410753 410754 410755 410756 410757 410758 410759 410760 410762 410763
#> [1111] 410764 410765 410766 410767 410768 410769 410770 410771 410772 410773
#> [1121] 410774 410775 410779 410782 410783 410784 410785 410786 410787 410788
#> [1131] 410791 410792 410793 410794 410795 410796 410797 410798 410799 410800
#> [1141] 410801 410803 410804 410805 410806 410807 410808 410809 410811 410812
#> [1151] 410813 410814 410815 410816 410818 410820 410821 410823 410825 410826
#> [1161] 410827 410828 410829 410830 410831 410832 410833 410834 410835 410836
#> [1171] 410837 410838 410839 410840 410841 410842 410843 410844 410845 410846
#> [1181] 410847 410848 410849 410850 410851 410852 410853 410855 410856 410857
#> [1191] 410858 410859 410860 410862 410863 410864 410865 410866 410867 410868
#> [1201] 410869 410870 410871 410872 410873 410874 410875 410876 410877 410878
#> [1211] 410879 410880 410881 410882 410883 410884 410885 410887 410888 410889
#> [1221] 410890 410891 410892 410894 410895 410897 410898 410899 410901 410906
#> [1231] 410907 410909 410912 410913 410914 410916 410917 410918 410919 410921
#> [1241] 410922 410923 410924 410925 410926 410927 410928 410930 410931 410932
#> [1251] 410933 410934 410935 410936 410937 410938 410942 410948 410949 410950
#> [1261] 410951 410952 410953 410954 410955 410956 410957 410958 410960 410961
#> [1271] 410962 410963 410964 410965 410966 410967 410968 410969 410970 410972
#> [1281] 410975 410976 410977 410978 410979 410980 410981 410982 410983 410984
#> [1291] 410985 410986 410987 410988 410989 410990 410991 410992 410993 410994
#> [1301] 410995 410996 410997 410998 410999 411000 411001 411681 411682 411683
#> [1311] 411685 411686 411687 411688 411689 411690 411691 411692 411693 411694
#> [1321] 411696 411697 411698 411699 411700 411701 411702 411703 411704 411705
#> [1331] 411706 411707 411708 411709 411710 411711 411712 411713 411714 411715
#> [1341] 411716 411717 411720 411721 411722 411724 411725 411730 411731 411732
#> [1351] 411733 411734 411735 411736 411738 411739 411740 411741 411742 411743
#> [1361] 411744 411745 411747 411748 411749 411752 411753 411758 411759 411761
#> [1371] 411762 411763 411764 411765 411766 411767 411768 411769 411770 411771
#> [1381] 411773 411774 411775 411776 411777 411780 411781 411782 411783 411784
#> [1391] 411785 411786 411787 411788 411790 411791 411792 411793 411794 411795
#> [1401] 411796 411797 411799 411801 411803 411805 411806 411807 411808 411809
#> [1411] 411810 411811 411813 411814 411815 411816 411818 411819 411822 411823
#> [1421] 411824 411825 411827 411828 411829 411832 411833 411834 411835 411836
#> [1431] 411837 411838 411839 411840 411841 411842 411843 411844 411846 411847
#> [1441] 411848 411849 411853 411855 411856 411857 412210 412211 412212 412213
#> [1451] 422362 422453 422483 423312 423313 423314 423334 423335 423336 423337
#> [1461] 423338 425236 425238 425244 425249 425254 425447 425692

# For many more options see ?ggplot2::theme