Map tree tags by status, showing four subquadrats per plot-page.

This function maps tree tags by status and outputs a list of maps that can be printed on a .pdf file. Each map shows four subquadrats within a quadrat. The symbols on the map represent the status of each tree -- not the status of each stem. Although you should likely provide data of only one or two censuses, maply_tag() will summarize the data to reduce overplotting. The data on the map summarizes the history of each stem accross all censuses provided. Each tag will appear in the map only once or twice: * A tag will appear once if it belongs to a tree which status was unique accross all censuses provided -- either "alive" or "dead". * A tag will appear twice if it belongs to a tree which status was "alive" in at least one census, and also "dead" in at least one other census. This feature avoids unintentional overplotting and makes interpreting the map easier.

maply_tag(vft, x_q = 20, x_sq = 5, y_q = 20, y_sq = 5,
  subquad_offset = NULL, bl = 1, br = 2, tr = 3, tl = 4,
  title_quad = "Site Name, YYYY. Quadrat:", show_page = TRUE,
  show_subquad = TRUE, point_shape = c(19, 4), point_size = 1.5,
  tag_size = 3, header = map_tag_header(), theme = theme_map_tag(),
  move_edge = 0)

Arguments

vft
A dataframe -- specifically, a ForestGEO ViewFullTable.

x_q, y_q
Size in meters of a quadrat's side. For ForestGEO sites, a common value is 20.

x_sq, y_sq
Size in meters of a subquadrat's side. For ForestGEO-CTFS sites, a common value is 5.

subquad_offset
NULL or -1. NULL defines the first column of subquadrats as 1. -1 defines the first column of subquadrats as 0.

subquad_offset = NULL    subquad_offset = -1
---------------------    -------------------
     14 24 34 44             04 14 24 34
     13 23 33 43             03 13 23 33
     12 22 32 42             02 12 22 32
     11 21 31 41             01 11 21 31

bl, br, tr, tl
Label each of the four maps of a quadrat. See paginate().

title_quad
A string to use as a title.

show_page
Logical; FALSE removes the page label from the map title.

show_subquad
Logical; FALSE removes subquadrat labels on each map.

point_shape
A vector of two numbers giving the shape of the points to plot (see possible shapes in the documentation of ?graphics::points(), under the section entitled 'pch' values).

point_size
A number giving points size. Passed to ggplot2::geom_point().

tag_size
A number giving tag size. Passed to ggrepel::geom_text_repel.

header
A string to use as a header (see headers).

theme
An object of class theme.

move_edge
A number to adjust the extension of the grid lines beyond the plot limits.

Value

A list of ggplots, where each element of the list is a map of tree tags by status, showing four subquadrats.

Acknowledgements

Useful ideas and guidance came from Suzanne Lao, Stuart Davis, Shameema Jafferjee Esufa, David Kenfack and Anudeep Singh. Andudeep Sinh also wrote the algorithm to calculate subquadrats.

See also

Examples

# NOT RUN {
library(fgeo.map)
library(fgeo.tool)
library(dplyr)

# Filtering the data to map -----------------------------------------------

# Filter the data you want. For example:
# (see ?vft_1quad)
want <- filter(
  vft_1quad,
  CensusID == 6,
  PlotID == 1
)
p <- maply_tag(want)
# Visualizing only the first plot of `p`
first(p)
# Printing all plots of `p` to .pdf, with parameters optimized for size letter
pdf("map.pdf", paper = "letter", height = 10.5, width = 8)
p
dev.off()

# Be careful if filtering by DBH: You may unintentionally remove dead trees.
# * Confirm this dataset has dead trees:
# (see `?vft_4quad`)
dead <- vft_4quad %>%
  add_status_tree(status_a = "alive", status_d = "dead") %>%
  pick_top(QuadratID) %>%
  filter(status_tree == "dead")
select(dead, Tag, Status, status_tree, DBH)
first(maply_tag(dead))
# * If you filter by `DBH`, you loose the dead trees becaue their `DBH = NA`
wrong <- filter(dead, DBH > 10)
wrong
# * The right way to do it is to explicietly inlcude rows where DBH = NA
right <- filter(dead, DBH > 10 | is.na(DBH))
right
first(maply_quad(right))

# Keeping dead trees with `is.na(DBH)` (e.g. tag 127885.d on the bottom right)
p <- filter(vft_4quad, DBH > 20 | is.na(DBH))
first(maply_tag(p))

# For more complex filtering, see also ?fgeo.tool::drop_twice_dead

# Customizing the maps ----------------------------------------------------

# Common tweaks

p <- maply_tag(vft_1quad, show_page = FALSE, show_subquad = FALSE)
first(p)

p <- maply_tag(
  vft_1quad,
  title_quad = "BCI 2012. Quadrat: ",
  bl = "bottom-left", br = "bottom-right", tr = "top-right", tl = "top-left",
  header = "Line 1: _________\nLine 2:\nLine 3:.....................",
  subquad_offset = -1,
  point_size = 3, point_shape = c(17, 6),
  tag_size = 2,
  move_edge = 0.5
)
first(p)

# Themes

library(ggplot2)

p <- maply_tag(vft_1quad, theme = theme_gray())
first(p)

# Tweaking the default theme of maply_tag()

small_tweak <- theme_map_tag(legend.position = "bottom")
p <- maply_tag(vft_1quad, theme = small_tweak)
first(p)

large_tweak <- theme(
  legend.position = "bottom",
  legend.title = element_blank(),
  legend.text = element_text(size = 8, colour = "red"),
  text = element_text(size = 11, face = "bold.italic", colour = "white"),
  plot.background = element_rect(fill = "black"),
  plot.margin = margin(2, 2, 2, 2, "cm"),
  strip.background = element_rect(fill = "darkgreen"),
  strip.text = element_text(colour = "white"),
  # make grid to dissapear by matching background colour
  panel.background = element_rect(fill = "lightgreen"),
  panel.grid.minor = element_line(colour = "black", linetype = "dotted"),
  panel.grid.major = element_line(colour = "black")
)
p <- maply_tag(vft_1quad, theme = large_tweak)
first(p)

# Quadrat and subquadrat dimensitons

# Adapting the dimensions of quadrat and subquadrat to the range of qx and qy
# Creating new data set with QX and QY ranging 0-100
smaller <- vft_1quad
n <- nrow(smaller)
smaller$QX <- sample(0:10, n, replace = TRUE)
smaller$QY <- sample(0:10, n, replace = TRUE)

p <- maply_tag(smaller, x_q = 10, x_sq = 2.5)
first(p)

# If limit-lines aren't visible, try extending the grid a little
p <- maply_tag(smaller, x_q = 10, x_sq = 2.5, move_edge = 0.25)[1]
first(p)
# }