Suffix tags of dead stems that map beyond the quadrat's edge.

Suffix tags of dead stems that map beyond the quadrat's edge.

suffix_tags_beyond_edge(x, .match, suffix, x_q = 20, y_q = x_q)

Arguments

x
A ViewFullTable.

.match
A character string indicating the matching value of the variable status to suffix.

suffix
A character string to suffix tree tags with.

x_q, y_q
Integer; the x and y size in meters of a quadrat.

Value

A modified version of x.

Examples

vft <- data.frame(
  QX = 21:22,
  QY = 21:22,
  Tag = c("01", "02"),
  Status = c("dead", "alive"),
  stringsAsFactors = FALSE
)
vft

#>   QX QY Tag Status
#> 1 21 21  01   dead
#> 2 22 22  02  alive

suffix_tags_beyond_edge(
  x = vft,
  .match = "dead",
  suffix = "_d",
  x_q = 20
)

#> One or more trees spillover.

#>   QX QY  Tag Status
#> 1 21 21 01_d   dead
#> 2 22 22   02  alive

suffix_tags_beyond_edge(
  x = vft,
  .match = "alive",
  suffix = "_a",
  x_q = 20
)

#> One or more trees spillover.

#>   QX QY  Tag Status
#> 1 21 21   01   dead
#> 2 22 22 02_a  alive